pyproject.toml
[project]
name = "agentic-drop-zones"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-aiplatform>=1.118.0",
]
